refactor(types): split blob.ts into entities/blob.ts
