+++ packages/types/src/api-constants/api-limits.ts
@@ -17,7 +17,7 @@ import {
   TWENTY_FIVE_MiB,
 } from "./time-constants.js";
 
-import type { BlobPurpose } from "../blob.js";
+import type { BlobPurpose } from "../entities/blob.js";
 import type { RotationItemStatus, RotationState } from "../key-rotation.js";
 
 // ── Pagination ───────────────────────────────────────────────────────
