+++ packages/types/src/entities/blob.ts
@@ -1,6 +1,6 @@
-import type { BlobId, ChecksumHex, SystemId } from "./ids.js";
-import type { UnixMillis } from "./timestamps.js";
-import type { Archived } from "./utility.js";
+import type { BlobId, ChecksumHex, SystemId } from "../ids.js";
+import type { UnixMillis } from "../timestamps.js";
+import type { Archived } from "../utility.js";
 
 /** Encryption tier for stored blobs: 1 = at-rest only, 2 = at-rest + per-bucket key. */
 export type EncryptionTier = 1 | 2;
